Initial commit
sonnetMakerMarkov.py and sonnetMakerTweets.py are the two scripts that you'll want to type into the command line.
intake.txt and iambicPentameter.txt are for storing information.
rhymes.py is to supply the scripts with some functions

Completed sonnets are added to completedSonnets.txt. Some sonnets are already in there.
As a bonus, rhyming couplets get added to neverEndingCoupletPoem.txt.